Pagamento feito
Falta adicionar na BD para o historico

+++ PROJETOESA/Models/PaymentModel.cs
@@ -2,7 +2,7 @@
 {
     public class PaymentModel
     {
-        public long price { get; set; }
+        public decimal price { get; set; }
         public required string currency { get; set; }
         public string Email { get; set; }
         public string CreditCard { get; set; }
